requirements.txt
clickhouse_connect==0.8.11
gradio==5.9.1
hydra-core==1.3.2
lightning==2.5.0.post0
numpy==2.2.1
omegaconf==2.3.0
pandas==2.2.3
polars==1.18.0
pytorch_lightning==2.5.0.post0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
